优化统一全局异常处理，添加页面请求注解

+++ src/main/java/com/example/springbootlearning/common/api/IErrorCode.java
@@ -13,7 +13,7 @@ public interface IErrorCode {
      *
      * @return long
      */
-    long getCode();
+    int getCode();
 
     /**
      * 获取消息
